+++ Changelog.md
@@ -6,6 +6,8 @@ This file only reflects the changes that are made in the the docker image. Pleas
 - set default value of `DB_TYPE` to `postgres`
 - added sample Kubernetes rc and service description files
 - expose `GITLAB_BACKUP_ARCHIVE_PERMISSIONS` parameter
+- gitlab: upgrade to CE v.8.0.0
+- added `GITLAB_SECRETS_DB_KEY_BASE` parameter
 
 **7.14.3**
 - gitlab: upgrade to CE v.7.14.3